Add a controller for server engines page
