+++ server_engines.php
@@ -1,31 +1,32 @@
 <?php
 /* vim: set expandtab sw=4 ts=4 sts=4: */
+
 /**
- * display list of server engines and additional information about them
+ * Handles server engines page.
  *
  * @package PhpMyAdmin
  */
 
-/**
- * requirements
- */
+namespace PMA;
+
+use PMA\libraries\controllers\server\ServerEnginesController;
+use PMA\libraries\Response;
+
 require_once 'libraries/common.inc.php';
 
-/**
- * Does the common work
- */
-require 'libraries/server_common.inc.php';
-require 'libraries/server_engines.lib.php';
+$container = libraries\di\Container::getDefaultContainer();
+$container->factory(
+    'PMA\libraries\controllers\server\ServerEnginesController'
+);
+$container->alias(
+    'ServerEnginesController',
+    'PMA\libraries\controllers\server\ServerEnginesController'
+);
+$container->set('PMA\libraries\Response', Response::getInstance());
+$container->alias('response', 'PMA\libraries\Response');
 
-/**
- * Displays the sub-page heading
- */
-$response = PMA\libraries\Response::getInstance();
-$response->addHTML(PMA_getHtmlForSubPageHeader('engines'));
-
-/**
- * start output
- */
-$response->addHTML(PMA_getHtmlForServerEngines());
-
-exit;
+/** @var ServerEnginesController $controller */
+$controller = $container->get(
+    'ServerEnginesController', array()
+);
+$controller->indexAction();
